swapped out unsigned types for NSUInteger type to get rid of warning message in XCode 4.
git-svn-id: svn://orca.physics.unc.edu/Orca@3843 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Cocoa Extensions/NSOutlineView+Extensions.m
@@ -20,7 +20,7 @@
 - (NSArray*)allSelectedItems {
     NSMutableArray *items = [NSMutableArray array];
 	NSIndexSet* selectedSet = [self selectedRowIndexes];
-	unsigned current_index = [selectedSet firstIndex];
+	NSUInteger current_index = [selectedSet firstIndex];
 	while (current_index != NSNotFound){
         if ([self itemAtRow:current_index]) [items addObject: [self itemAtRow:current_index]];
 		current_index = [selectedSet indexGreaterThanIndex: current_index];
